+++ lib/haytni_web/controllers/authenticable/session_controller.ex
@@ -4,9 +4,9 @@ defmodule HaytniWeb.Authenticable.SessionController do
   use HaytniWeb, :controller
   use HaytniWeb.Helpers, {Haytni.AuthenticablePlugin, :with_current_user}
 
-  defp redirect(conn) do
+  defp redirect_to(conn, path) do
     conn
-    |> redirect(to: "/")
+    |> redirect(to: path)
     |> halt()
   end
 
@@ -20,15 +20,14 @@ defmodule HaytniWeb.Authenticable.SessionController do
     render_new(conn, Haytni.AuthenticablePlugin.session_changeset(config))
   end
 
-  def new(conn, _params, _current_user, _module, _config) do
-    redirect(conn)
+  def new(conn, _params, _current_user, _module, config) do
+    redirect_to(conn, config.sign_in_return_path)
   end
 
   def create(conn, %{"session" => session_params}, nil, module, config) do
     case Haytni.AuthenticablePlugin.authenticate(conn, module, config, session_params) do
       {:ok, conn} ->
-        conn
-        |> redirect()
+        redirect_to(conn, config.sign_in_return_path)
       {:error, changeset = %Ecto.Changeset{}} ->
         conn
         |> HaytniWeb.Helpers.set_suspicious_activity()
@@ -36,17 +35,17 @@ defmodule HaytniWeb.Authenticable.SessionController do
     end
   end
 
-  def create(conn, _params, _current_user, _module, _config) do
-    redirect(conn)
+  def create(conn, _params, _current_user, _module, config) do
+    redirect_to(conn, config.sign_in_return_path)
   end
 
-  def delete(conn, _params, nil, _module, _config) do
-    redirect(conn)
+  def delete(conn, _params, nil, _module, config) do
+    redirect_to(conn, config.sign_out_return_path)
   end
 
-  def delete(conn, _params, _current_user, module, _config) do
+  def delete(conn, _params, _current_user, module, config) do
     conn
     |> Haytni.logout(module)
-    |> redirect()
+    |> redirect_to(config.sign_out_return_path)
   end
 end