+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   + [LiveViewPlugin] `Phoenix.LiveView.get_connect_info/1` => `Phoenix.LiveView.get_connect_info/2`
 - [Rememberable] fixed wrong cookie max-age value (expiration)
 - fixed missing SQL negation in `Haytni.Token.purge_expired_tokens/1`, removing valid tokens instead of expired
+- [Authenticable] added `sign_in_return_path` and `sign_out_return_path` options
 
 
 ## 0.7.0 (2022-04-10)